feat(lld): make icon wrapper a new component and support light mode
