+++ libs/ui/packages/react/src/components/asorted/Icon/index.tsx
@@ -3,3 +3,4 @@ import BoxedIcon, { IconBox } from "./BoxedIcon";
 export { BoxedIcon, IconBox };
 export type { Props } from "./Icon";
 export { default as CryptoIcon } from "./CryptoIcon";
+export { default as NotificationIcon } from "./NotificationIcon";
