+++ libs/ui/packages/react/src/components/asorted/index.ts
@@ -1,3 +1,3 @@
 export { default as Divider } from "./Divider";
-export { default as Icon, IconBox, BoxedIcon, CryptoIcon } from "./Icon";
+export { default as Icon, IconBox, BoxedIcon, CryptoIcon, NotificationIcon } from "./Icon";
 export { default as Text } from "./Text";
